requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
dotenv==0.9.9
python-dotenv==1.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-cloud-pubsub==2.29.0
ruff==0.11.4
kafka-python==2.1.5
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
# psycopg
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
# faiss
faiss-cpu==1.10.0
# celery
celery==5.5.1
# redis
redis==5.2.1